@@ -8,11 +8,15 @@
     "authors": [{
         "name": "Jordi Boggiano",
         "email": "j.boggiano@seld.be",
-        "homepage": "http://seld.be"
+        "homepage": "https://seld.be"
     }, {
         "name": "Nils Adermann",
         "email": "naderman@naderman.de",
         "homepage": "http://www.naderman.de"
+    }, {
+        "name": "Rob Bast",
+        "email": "rob.bast@gmail.com",
+        "homepage": "https://robbast.nl"
     }],
     "support": {
         "issues": "https://github.com/composer/satis",
@@ -22,19 +26,19 @@
     },
     "bin": ["bin/satis"],
     "require": {
-        "php": "^7.0",
-        "composer/composer": "^1.4",
-        "twig/twig": "^1.7 || ^2.0",
-        "symfony/console": "^2.1 || ^3.0.4 || ^4.0"
+        "php": "^7.3||^8.0",
+        "composer/composer": "^1.9",
+        "symfony/console": "^4.4||^5.0",
+        "twig/twig": "^2.12||^3.0"
     },
     "autoload": {
         "psr-4": { "Composer\\Satis\\": "src" }
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.4",
-        "phpunit/phpunit-mock-objects": "^4.0",
-        "mikey179/vfsstream": "^1.6",
-        "friendsofphp/php-cs-fixer": "^2.11"
+        "ext-json": "*",
+        "friendsofphp/php-cs-fixer": "^2.15",
+        "mikey179/vfsstream": "^1.6.7",
+        "phpunit/phpunit": "^8.2||^9.0"
     },
     "autoload-dev": {
         "psr-4": { "Composer\\Satis\\": "tests" }
@@ -43,11 +47,6 @@
         "build": [
             "@php bin/satis build satis.json build"
         ]
-    },
-    "config": {
-        "platform": {
-            "php": "7.0"
-        }
     },
     "extra": {
         "branch-alias": {
