@@ -16,25 +16,13 @@ namespace Composer\Satis\Builder;
 use Composer\Package\PackageInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Builds the archives of the repository.
- *
- * @author James Hautot <james@rezo.net>
- */
 class ArchiveBuilderHelper
 {
     /** @var OutputInterface The output Interface. */
     private $output;
-
     /** @var array The 'archive' part of a configuration file. */
     private $archiveConfig;
 
-    /**
-     * Helper Constructor.
-     *
-     * @param OutputInterface $output        The output Interface
-     * @param array           $archiveConfig The 'archive' part of a configuration file
-     */
     public function __construct(OutputInterface $output, array $archiveConfig)
     {
         $this->output = $output;
@@ -44,14 +32,7 @@ class ArchiveBuilderHelper
         $this->archiveConfig['blacklist'] = (array) ($archiveConfig['blacklist'] ?? []);
     }
 
-    /**
-     * Gets the directory where to dump archives.
-     *
-     * @param string $outputDir The directory where to build
-     *
-     * @return string $directory The directory where to dump archives
-     */
-    public function getDirectory($outputDir)
+    public function getDirectory(string $outputDir): string
     {
         if (isset($this->archiveConfig['absolute-directory'])) {
             $directory = $this->archiveConfig['absolute-directory'];
@@ -62,14 +43,7 @@ class ArchiveBuilderHelper
         return $directory;
     }
 
-    /**
-     * Tells if a package has to be dumped or not.
-     *
-     * @param PackageInterface $package The package to be dumped
-     *
-     * @return bool false if the package has to be dumped
-     */
-    public function isSkippable(PackageInterface $package)
+    public function isSkippable(PackageInterface $package): bool
     {
         if ('metapackage' === $package->getType()) {
             return true;
@@ -100,17 +74,7 @@ class ArchiveBuilderHelper
         return false;
     }
 
-    /**
-     * Check if any of the names is in the list.
-     *
-     * Any * in the list is treated as a wildcard.
-     *
-     * @param array $names Names to check
-     * @param array $list  List to check the names against
-     *
-     * @return bool true if any of the names is in the list
-     */
-    protected function isOneOfNamesInList(array $names, array $list)
+    protected function isOneOfNamesInList(array $names, array $list): bool
     {
         $patterns = $this->convertListToRegexPatterns($list);
 
@@ -123,15 +87,7 @@ class ArchiveBuilderHelper
         return false;
     }
 
-    /**
-     * Check if the name matches any of the patterns.
-     *
-     * @param string $name     Name to check
-     * @param array  $patterns Patterns to check the name against
-     *
-     * @return bool true if the name matches any of the patterns
-     */
-    protected function doesNameMatchOneOfPatterns($name, array $patterns)
+    protected function doesNameMatchOneOfPatterns(string $name, array $patterns): bool
     {
         foreach ($patterns as $pattern) {
             if (preg_match($pattern, $name)) {
@@ -142,16 +98,7 @@ class ArchiveBuilderHelper
         return false;
     }
 
-    /**
-     * Convert a list to regex patterns for use in preg_ functions.
-     *
-     * Any * is replaced with .* and the rest is escaped.
-     *
-     * @param array $list List to convert to patterns
-     *
-     * @return array array of patterns
-     */
-    protected function convertListToRegexPatterns(array $list)
+    protected function convertListToRegexPatterns(array $list): array
     {
         $patterns = [];
 
