@@ -1,17 +1,14 @@
-import $ from 'jquery';
-// import 'bootstrap';
-// import 'popper.js';
-import 'bootstrap/js/dist/collapse';
-import DateDistance from './App/DateDistance';
-import PackageFilter from './App/PackageFilter';
+import $ from 'jquery'
+import 'bootstrap/js/dist/collapse'
+import DateDistance from './App/DateDistance'
+import PackageFilter from './App/PackageFilter'
 
 function updateTimeElements () {
-    DateDistance.calculate('time')
+  DateDistance.calculate('time')
 };
 
-$(document).ready(function() {
-    new PackageFilter('input#search', '#package-list', '.card');
-
-    updateTimeElements();
-    window.setInterval(updateTimeElements, 5000);
+$(function() {
+  new PackageFilter('input#search', '#package-list', '.card')
+  updateTimeElements()
+  window.setInterval(updateTimeElements, 5000)
 });
